Add files via upload
While our exploration of the data, there were few missing values in the data set, which we need to be solve before the data is ready for a good model. 

There are missing values in some variables. We should estimate those values wisely depending on the amount of missing values and the expected importance of variables.

In addition to these problems with numerical fields, there are non-numerical fields i.e. Gender, Property_Area, Married, Education and Dependents to see, if they contain any useful information and thus fill the missing values by using pandas techniques in python. Like boolean indexing, apply function.