Add files via upload
While our exploration of the data, there were few missing values in the data set, which needs to be solved before the data is ready for a good model. Problems, we are already aware of:

Missing values in some variables. Estimate those values depending on the amount of missing values and the expected importance of variables.

In addition to these problems with numerical fields, we need to consider the non-numerical fields too i.e. Gender, Property_Area, Married, Education and Dependents to see, if they contain any useful information likewise by using pandas techniques impute the missing values.